Store Support/Projects/VET_Dashboard/requirements.txt
google-cloud-bigquery==3.18.0
google-auth==2.28.0
google-cloud-core==2.4.1
python-dotenv==1.0.0
python-pptx==0.6.23
Pillow==10.1.0
Flask==2.3.0
Flask-CORS==4.0.0